requirements.txt
Django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

pytest==7.4.3
black==24.1.1
flake8==7.0.0 